r2769/sofa-dev : FIX: remove some warnings ADD: add log in Mesh and SphereModel when a wrong filename is used
 //partial sync (6/7 changes)//

+++ modules/sofa/component/forcefield/TriangularFEMForceField.cpp
@@ -1073,7 +1073,7 @@ void TriangularFEMForceField<DataTypes>::draw()
 
             float meanEV = (float)((vertexInfo[a].sumEigenValues + vertexInfo[b].sumEigenValues + vertexInfo[c].sumEigenValues) / 3.0);
 
-            glColor4f(meanEV / max, 1- meanEV / max, 0, 1);
+            glColor4f((float)(meanEV / max), (float)(1- meanEV / max), 0.0f, 1.0f);
             helper::gl::glVertexT(x[a]);
             helper::gl::glVertexT(x[b]);
             helper::gl::glVertexT(x[c]);
